adding invoices option

+++ controllers/admin.js
@@ -1,6 +1,7 @@
 const { validationResult } = require('express-validator');
 
 const mongoose = require('mongoose');
+const fileHelper = require('../util/file');
 
 const Product = require('../models/product');
 
@@ -18,10 +19,25 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.body.image;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
     const errors = validationResult(req);
+    if (!image) {
+        return res.status(422).render('admin/edit-product', {
+            pageTitle: 'Add product',
+            path: '/admin/add-product',
+            editing: false,
+            hasError: true,
+            product: {
+                title: title,
+                price: price,
+                description: description
+            },
+            errorMessage: 'Attached file is not an image.',
+            validationErrors: []
+        });
+    }
 
     if (!errors.isEmpty()) {
         console.log(errors);
@@ -40,6 +56,8 @@ exports.postAddProduct = (req, res, next) => {
             validationErrors: errors.array()
         });
     }
+
+    const imageUrl = image.path;
 
     const product = new Product({
         title: title,
@@ -109,10 +127,12 @@ exports.getAEditProduct = (req, res, next) => {
 exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
     const updatedPrice = req.body.price;
     const updatedDescription = req.body.description;
     const errors = validationResult(req);
+
+
 
     if (!errors.isEmpty()) {
         console.log(errors);
@@ -124,7 +144,6 @@ exports.postEditProduct = (req, res, next) => {
             product: {
                 _id: prodId,
                 title: updatedTitle,
-                imageUrl: updatedImageUrl,
                 price: updatedPrice,
                 description: updatedDescription
             },
@@ -132,7 +151,6 @@ exports.postEditProduct = (req, res, next) => {
             validationErrors: errors.array()
         });
     }
-
     Product.findById(prodId)
         .then(product => {
             if (product.userId.toString() !== req.user._id.toString()) {
@@ -141,7 +159,10 @@ exports.postEditProduct = (req, res, next) => {
             product.title = updatedTitle;
             product.price = updatedPrice;
             product.description = updatedDescription;
-            product.imageUrl = updatedImageUrl;
+            if (image) {
+                fileHelper.deleteFile(product.imageUrl);
+                product.imageUrl = image.path;
+            }
             return product.save()
                 .then(result => {
                     console.log('Product Updated');
@@ -153,17 +174,20 @@ exports.postEditProduct = (req, res, next) => {
             error.httpStatusCode = 500;
             return next(error);
         });
-
-
 };
 
 exports.postDeleteProduct = (req, res, data) => {
     const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
-        .then(() => {
-            console.log('Product destroyed!');
-            res.redirect('/admin/products');
-        })
+    Product.findById(prodId).then(product => {
+        if (!product) {
+            return next(new Error('Product not found'));
+        }
+        fileHelper.deleteFile(product.imageUrl);
+        return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    }).then(() => {
+        console.log('Product destroyed!');
+        res.redirect('/admin/products');
+    })
         .catch(err => {
             const error = new Error(err);
             error.httpStatusCode = 500;
